tests: start and restructure test to use db_session fixture in multiple files
